requirements.txt
pandas
numpy
seaborn
matplotlib
plotly
nbformat>=4.2.0
streamlit==1.33.0
Pillow